created functions ssmp_recv_color_start and ssmp_recv_from_test.
The former is a reentry version that receives as a parameter the core sequences from which the testing for incoming msgs should start. It also returns the seq of the core from which the receive happened.

The latter tets if there is a incoming message without changing the flag.

FIX: forgotten # bef include